+++ src/main/java/io/github/ryuu/adventurecraft/items/ItemPowerGlove.java
@@ -1,9 +1,13 @@
 package io.github.ryuu.adventurecraft.items;
 
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
-import org.lwjgl.input.Keyboard;
 import io.github.ryuu.adventurecraft.blocks.Blocks;
+import net.minecraft.entity.FallingTile;
+import net.minecraft.entity.player.Player;
+import net.minecraft.item.ItemInstance;
+import net.minecraft.item.ItemType;
+import net.minecraft.level.Level;
+import net.minecraft.tile.Tile;
+import org.lwjgl.input.Keyboard;
 
 public class ItemPowerGlove extends ItemType {
 
@@ -30,7 +34,7 @@ public class ItemPowerGlove extends ItemType {
         if (level.getTileId(x, y, z) != Blocks.pushableBlock.id) {
             return false;
         }
-        if (Keyboard.isKeyDown((int) 29) || Keyboard.isKeyDown((int) 157)) {
+        if (Keyboard.isKeyDown(29) || Keyboard.isKeyDown(157)) {
             xOffset *= -1;
             zOffset *= -1;
         }
