+++ src/main/java/io/github/ryuu/adventurecraft/mixin/client/gui/screen/MixinDeathScreen.java
@@ -1,11 +1,11 @@
 package io.github.ryuu.adventurecraft.mixin.client.gui.screen;
 
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
+import net.minecraft.client.gui.Screen;
+import net.minecraft.client.gui.screen.TitleScreen;
+import net.minecraft.client.gui.widgets.Button;
 import org.lwjgl.opengl.GL11;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;
-import org.spongepowered.asm.mixin.Shadow;
 
 @Mixin(DeathScreen.class)
 public class MixinDeathScreen extends Screen {
@@ -36,7 +36,7 @@ public class MixinDeathScreen extends Screen {
     public void render(int mouseX, int mouseY, float delta) {
         this.fillGradient(0, 0, this.width, this.height, 0x60500000, -1602211792);
         GL11.glPushMatrix();
-        GL11.glScalef((float) 2.0f, (float) 2.0f, (float) 2.0f);
+        GL11.glScalef(2.0f, 2.0f, 2.0f);
         this.drawTextWithShadowCentred(this.textManager, "Game over!", this.width / 2 / 2, 30, 0xFFFFFF);
         GL11.glPopMatrix();
         super.render(mouseX, mouseY, delta);
