+++ src/main/java/io/github/ryuu/adventurecraft/mixin/entity/MixinFallingTile.java
@@ -1,7 +1,10 @@
 package io.github.ryuu.adventurecraft.mixin.entity;
 
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
+import net.minecraft.entity.Entity;
+import net.minecraft.level.Level;
+import net.minecraft.tile.SandTile;
+import net.minecraft.util.io.CompoundTag;
+import net.minecraft.util.maths.MathsHelper;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;
 import org.spongepowered.asm.mixin.Shadow;
@@ -59,20 +62,20 @@ public class MixinFallingTile extends Entity {
         this.prevY = this.y;
         this.prevZ = this.z;
         ++this.field_848;
-        this.velocityY -= (double) 0.04f;
+        this.velocityY -= 0.04f;
         this.move(this.velocityX, this.velocityY, this.velocityZ);
-        this.velocityX *= (double) 0.98f;
-        this.velocityY *= (double) 0.98f;
-        this.velocityZ *= (double) 0.98f;
+        this.velocityX *= 0.98f;
+        this.velocityY *= 0.98f;
+        this.velocityZ *= 0.98f;
         int i = MathsHelper.floor(this.x);
         int j = MathsHelper.floor(this.y);
         int k = MathsHelper.floor(this.z);
         if (this.level.getTileId(i, j, k) == this.tile) {
             this.level.setTile(i, j, k, 0);
         }
-        if (this.onGround && Math.abs((double) this.velocityX) < 0.01 && Math.abs((double) this.velocityZ) < 0.01) {
-            this.velocityX *= (double) 0.7f;
-            this.velocityZ *= (double) 0.7f;
+        if (this.onGround && Math.abs(this.velocityX) < 0.01 && Math.abs(this.velocityZ) < 0.01) {
+            this.velocityX *= 0.7f;
+            this.velocityZ *= 0.7f;
             this.velocityY *= -0.5;
             if (!SandTile.method_435(this.level, i, j - 1, k)) {
                 this.remove();
@@ -88,11 +91,11 @@ public class MixinFallingTile extends Entity {
             this.dropItem(this.tile, 1);
             this.remove();
         }
-        if (Math.abs((double) (this.x - this.startX)) >= 1.0) {
+        if (Math.abs(this.x - this.startX) >= 1.0) {
             this.velocityX = 0.0;
             this.setPosition((double) i + 0.5, this.y, this.z);
         }
-        if (Math.abs((double) (this.z - this.startZ)) >= 1.0) {
+        if (Math.abs(this.z - this.startZ) >= 1.0) {
             this.velocityZ = 0.0;
             this.setPosition(this.x, this.y, (double) k + 0.5);
         }
