+++ src/main/java/io/github/ryuu/adventurecraft/mixin/entity/monster/MixinSkeleton.java
@@ -1,7 +1,14 @@
 package io.github.ryuu.adventurecraft.mixin.entity.monster;
 
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.ItemEntity;
+import net.minecraft.entity.monster.Monster;
+import net.minecraft.entity.projectile.Arrow;
+import net.minecraft.item.ItemInstance;
+import net.minecraft.item.ItemType;
+import net.minecraft.level.Level;
+import net.minecraft.util.io.CompoundTag;
+import net.minecraft.util.maths.MathsHelper;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;
 import org.spongepowered.asm.mixin.Shadow;
@@ -42,7 +49,7 @@ public class MixinSkeleton extends Monster {
             double d1 = entity.z - this.z;
             if (this.attackTime == 0) {
                 Arrow entityarrow = new Arrow(this.level, this, this.attackDamage);
-                entityarrow.y += (double) 1.4f;
+                entityarrow.y += 1.4f;
                 double d2 = entity.y + (double) entity.getStandingEyeHeight() - (double) 0.2f - entityarrow.y;
                 float f1 = MathsHelper.sqrt(d * d + d1 * d1) * 0.2f;
                 this.level.playSound(this, "random.bow", 1.0f, 1.0f / (this.rand.nextFloat() * 0.4f + 0.8f));
@@ -50,7 +57,7 @@ public class MixinSkeleton extends Monster {
                 entityarrow.method_1291(d, d2 + (double) f1, d1, 0.6f, 12.0f);
                 this.attackTime = 30;
             }
-            this.yaw = (float) (Math.atan2((double) d1, (double) d) * 180.0 / 3.1415927410125732) - 90.0f;
+            this.yaw = (float) (Math.atan2(d1, d) * 180.0 / 3.1415927410125732) - 90.0f;
             this.field_663 = true;
         }
     }
@@ -84,8 +91,7 @@ public class MixinSkeleton extends Monster {
             i = this.rand.nextInt(3) + 1;
             for (int j = 0; j < i; ++j) {
                 ItemEntity item = this.dropItem(this.getMobDrops(), 1);
-                if (this.getMobDrops() == ItemType.arrow.id)
-                    continue;
+                if (this.getMobDrops() == ItemType.arrow.id) continue;
                 item.item.count = 3;
             }
         }
