+++ src/main/java/io/github/ryuu/adventurecraft/rendering/RenderBipedScaledScripted.java
@@ -1,9 +1,10 @@
 package io.github.ryuu.adventurecraft.rendering;
 
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
-import org.lwjgl.opengl.GL11;
 import io.github.ryuu.adventurecraft.entities.EntityLivingScript;
+import net.minecraft.client.render.entity.BipedEntityRenderer;
+import net.minecraft.client.render.entity.model.BipedModel;
+import net.minecraft.entity.LivingEntity;
+import org.lwjgl.opengl.GL11;
 
 public class RenderBipedScaledScripted extends BipedEntityRenderer {
 
@@ -17,6 +18,6 @@ public class RenderBipedScaledScripted extends BipedEntityRenderer {
         float width = (1.0f - f) * e.prevWidth + f * e.width;
         float height = (1.0f - f) * e.prevHeight + f * e.height;
         this.field_2678 = (width /= 0.6f) * 0.5f;
-        GL11.glScalef((float) width, (float) (height / 1.8f), (float) width);
+        GL11.glScalef(width, height / 1.8f, width);
     }
 }
